requirements.txt
aiohappyeyeballs==2.3.4
aiohttp==3.10.1
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
attrs==24.1.0
certifi==2024.7.4
charset-normalizer==3.3.2
dataclasses-json==0.6.7
decorator==5.1.1
distro==1.9.0
frozenlist==1.4.1
graphdatascience==1.11
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.2.12
langchain-community==0.2.11
langchain-core==0.2.28
langchain-experimental==0.0.64
langchain-openai==0.1.20
langchain-text-splitters==0.2.2
langgraph==0.1.19
langsmith==0.1.96
marshmallow==3.21.3
multidict==6.0.5
multimethod==1.12
mypy-extensions==1.0.0
neo4j==5.23.1
numpy==1.26.4
openai==1.39.0
orjson==3.10.6
packaging==24.1
pandas==2.2.2
py==1.11.0
pyarrow==16.1.0
pydantic==2.8.2
pydantic_core==2.20.1
PyMuPDF==1.24.9
PyMuPDFb==1.24.9
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
regex==2024.7.24
requests==2.32.3
retry==0.9.2
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.31
tenacity==8.5.0
textdistance==4.6.3
tiktoken==0.7.0
tqdm==4.66.5
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
yarl==1.9.4
